Fixed scrobbling the default Music app on Android 2.3 (Gingerbread). We can no longer access the MediaPlaybackService because it was marked as non-exported here: http://android.git.kernel.org/?p=platform/packages/apps/Music.git;a=commit;h=87a4cb677a3656ddf770602c196afb7b074371a9 but the good news is that we don't need to because we can now get all the information we need from the broadcasts because they are now sent when playback stops at the end of a playlist: http://android.git.kernel.org/?p=platform/packages/apps/Music.git;a=commit;h=ab893d7a8067b9c45eda455b38f3b064e82dcb91 As the old saying goes, "Google never closes a door without opening a window."

+++ src/net/jjc1138/android/scrobbler/ScrobblerService.java
@@ -75,7 +75,7 @@ class Track implements Serializable {
 			}
 		}
 		
-		id = i.getIntExtra("id", -1);
+		id = i.getLongExtra("id", -1);
 		
 		if (id != -1) {
 			final String[] columns = new String[] {
@@ -223,7 +223,7 @@ class Track implements Serializable {
 
 	@Override
 	public int hashCode() {
-		return id;
+		return (int) id;
 	}
 
 	@Override
@@ -232,7 +232,7 @@ class Track implements Serializable {
 		return s + artist + " - " + track;
 	}
 
-	private int id;
+	private long id;
 
 	private String artist;
 	private String track;
@@ -653,7 +653,7 @@ public class ScrobblerService extends Service {
 	private void handleIntent(Intent intent) {
 		Log.v(LOG_TAG, "Status: " +
 			((intent.getBooleanExtra("playing", false) ? "playing" : "stopped")
-			+ " track " + intent.getIntExtra("id", -1)));
+			+ " track " + intent.getLongExtra("id", -1)));
 		
 		if (!prefs.getBoolean("enable", true)) {
 			return;
